requirements.txt
setuptools>=65.5.1
wheel
Mako>=1.2.2
alembic==1.4.0
azure-common==1.1.24
azure-nspkg==3.0.2
azure-storage==0.36.0
certifi>=2019.11.28
cffi>=1.14.6
chardet>=3.0.4
Click==7.1.2
cryptography>=2.8
Flask==2.0.3
Flask-Login==0.5.0
Flask-Session==0.4.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.14.3
idna>=2.8
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
msal==1.13.0
pycparser>=2.20
pyodbc==5.2.0
pyopenssl>=19.1.0
python-dateutil>=2.8.1
python-editor==1.0.4
requests>=2.22.0
six>=1.14.0
SQLAlchemy==1.3.13
urllib3>=1.25.8
Werkzeug==2.0.3
WTForms==2.2.1
